address PR #23 comments
